atsay: Enable fast and dry-run modes
Full generation, complete with soft-mute support, is expensive and
therefore slow, especially on a weak computer. Fast mode avoids the soft
muting but, other than that, generates audio as normal. Dry run is
effectively just a syntax checker, audio is not generated and the durations
of speech fragments is just estimated